[Feat]: Create Comment

+++ shareFood/home/views.py
@@ -16,22 +16,35 @@ class IsOwnerOrReadOnly(permissions.BasePermission):
     def has_object_permission(self, request, view, obj):
         if request.method in permissions.SAFE_METHODS:
             return True
-        # return obj.writer == request.user
+    
         # 로그인하지 않은 사용자에게 GET 요청을 허용합니다.
         if request.method == 'GET' and not request.user.is_authenticated:
             return True
-        return obj.writer == request.user
+        return obj.user == request.user
+
+# class IsCommentAuthorOrReadOnly(permissions.BasePermission):
+#     def has_permission(self, request, view):
+#         # Allow GET, HEAD, or OPTIONS requests without checking further
+#         if request.method in permissions.SAFE_METHODS:
+#             return True
+#         return request.user and request.user.is_authenticated
+
+#     def has_object_permission(self, request, view, obj):
+#         # Allow the author of the comment to edit or delete it
+#         return obj.user == request.user
 
 class IsCommentAuthorOrReadOnly(permissions.BasePermission):
-    def has_permission(self, request, view):
-        # Allow GET, HEAD, or OPTIONS requests without checking further
+    def has_object_permission(self, request, view, obj):
+        # 읽기 권한은 모든 요청에 허용
         if request.method in permissions.SAFE_METHODS:
             return True
-        return request.user and request.user.is_authenticated
 
-    def has_object_permission(self, request, view, obj):
-        # Allow the author of the comment to edit or delete it
+        # 댓글을 수정하거나 삭제하려는 사용자가 댓글의 작성자와 같은 경우에만 허용
         return obj.user == request.user
+
+
+
+
 
 class DeliveryViewSet(viewsets.ModelViewSet):
     queryset = Delivery.objects.all().order_by('-id')
@@ -40,7 +53,7 @@ class DeliveryViewSet(viewsets.ModelViewSet):
 
     #글작성
     def perform_create(self, serializer):
-        serializer.save(writer=self.request.user)
+        serializer.save(user=self.request.user)
     
     #상태변경
     @action(detail=True, methods=['PATCH'], permission_classes=[IsOwnerOrReadOnly])
@@ -50,7 +63,7 @@ class DeliveryViewSet(viewsets.ModelViewSet):
         except Delivery.DoesNotExist:
             return Response({'error': '게시글을 찾을 수 없음.'}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user == delivery.writer:
+        if request.user == delivery.user:
             delivery.is_completed = True  # 구매 확정을 True로 설정
             delivery.save()
             return Response({'status': '구매가 확정되었습니다.'}, status=status.HTTP_200_OK)
@@ -64,7 +77,7 @@ class DeliveryViewSet(viewsets.ModelViewSet):
         except Delivery.DoesNotExist:
             return Response({'error': '게시글을 찾을 수 없음.'}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user == delivery.writer:
+        if request.user == delivery.user:
             delivery.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         else:
@@ -129,7 +142,7 @@ class GroceryViewSet(viewsets.ModelViewSet):
         except Grocery.DoesNotExist:
             return Response({'error': '게시글을 찾을 수 없음.'}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user == grocery.writer:
+        if request.user == grocery.user:
             grocery.is_completed = True  # 구매 확정을 True로 설정
             grocery.save()
             return Response({'status': '구매가 확정되었습니다.'}, status=status.HTTP_200_OK)
@@ -143,27 +156,13 @@ class GroceryViewSet(viewsets.ModelViewSet):
         except Grocery.DoesNotExist:
             return Response({'error': '게시글을 찾을 수 없음.'}, status=status.HTTP_404_NOT_FOUND)
 
-        if request.user == grocery.writer:
+        if request.user == grocery.user:
             grocery.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         else:
             return Response({'error': '권한이 없습니다.(작성자가 아님)'}, status=status.HTTP_403_FORBIDDEN)
         
-
-# class GroceryCommentViewSet(viewsets.ModelViewSet):
-#     queryset = GroceryComment.objects.all()
-#     serializer_class = GroceryCommentSerializer
-#     permission_classes = [IsCommentAuthorOrReadOnly]
-
-#     def perform_create(self, serializer):
-#         delivery_id = self.kwargs['id']  # 이 부분에서 URL에서 delivery_id를 가져옵니다.
-#         delivery = Delivery.objects.get(id=delivery_id)
-#         serializer.save(post=delivery, user=self.request.user)
-    
-#     # create 메서드 추가
-#     def create(self, request, *args, **kwargs):
-#         return super().create(request, *args, **kwargs)
-
+#### grocery 댓글 기능 ####
 
 class GroceryCommentView(APIView):   # 댓글 리스트
     permission_classes = [IsCommentAuthorOrReadOnly]
@@ -181,19 +180,80 @@ class GroceryCommentView(APIView):   # 댓글 리스트
         else:
             return Response(serializer.error_messages, status=status.HTTP_400_BAD_REQUEST)
 
-class GroceryCommentDetailView(APIView):   # 댓글 상세
+
+class GroceryCommentDetailView(APIView):
     permission_classes = [IsCommentAuthorOrReadOnly]
     
     def patch(self, request, post_id, comment_id):
         comment = get_object_or_404(GroceryComment, pk=comment_id)
-        serializer = GroceryCommentSerializer(comment, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save(user=request.user, post_id=post_id)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        
+        # 댓글 작성자만 수정할 수 있도록 검사
+        if comment.user == request.user:
+            serializer = GroceryCommentSerializer(comment, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save(user=request.user, post_id=post_id)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.error_messages, status=status.HTTP_400_BAD_REQUEST)
-    
+            return Response({'message': '댓글 작성자만 수정할 수 있습니다.'}, status=status.HTTP_403_FORBIDDEN)
+
     def delete(self, request, post_id, comment_id):
         comment = get_object_or_404(GroceryComment, pk=comment_id)
-        comment.delete()
-        return Response({'message': '삭제되었습니다.'}, status=status.HTTP_204_NO_CONTENT)
+
+        # 댓글 작성자만 삭제할 수 있도록 검사
+        if comment.user == request.user:
+            comment.delete()
+            return Response({'message': '삭제되었습니다.'}, status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'message': '댓글 작성자만 삭제할 수 있습니다.'}, status=status.HTTP_403_FORBIDDEN)
+
+
+
+
+
+#### Delivery 댓글 기능  ####
+
+class DeliveryCommentView(APIView):   # 댓글 리스트
+    permission_classes = [IsCommentAuthorOrReadOnly]
+
+    def get(self, request, post_id):
+        comment = DeliveryComment.objects.filter(post=post_id)
+        serializer = DeliveryCommentSerializer(comment, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request, post_id):
+        serializer = DeliveryCommentSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(user=request.user, post_id=post_id)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.error_messages, status=status.HTTP_400_BAD_REQUEST)
+
+
+class DeliveryCommentDetailView(APIView):
+    permission_classes = [IsCommentAuthorOrReadOnly]
+    
+    def patch(self, request, post_id, comment_id):
+        comment = get_object_or_404(DeliveryComment, pk=comment_id)
+        
+        # 댓글 작성자만 수정할 수 있도록 검사
+        if comment.user == request.user:
+            serializer = DeliveryCommentSerializer(comment, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save(user=request.user, post_id=post_id)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({'message': '댓글 작성자만 수정할 수 있습니다.'}, status=status.HTTP_403_FORBIDDEN)
+
+    def delete(self, request, post_id, comment_id):
+        comment = get_object_or_404(DeliveryComment, pk=comment_id)
+
+        # 댓글 작성자만 삭제할 수 있도록 검사
+        if comment.user == request.user:
+            comment.delete()
+            return Response({'message': '삭제되었습니다.'}, status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'message': '댓글 작성자만 삭제할 수 있습니다.'}, status=status.HTTP_403_FORBIDDEN)
